move mapping method to extension class
